+++ src/CommonController.php
@@ -5,7 +5,9 @@ namespace tsaotai\addons;
 
 /**
  * 插件通用基础控制器（向后兼容）
- * 请使用 BaseController 替代
+ * 
+ * @deprecated 请使用 BaseController 替代
+ * @see \tsaotai\addons\BaseController
  */
 abstract class CommonController extends BaseController
 {